Fix MBR bitfields

+++ include/kernel/mbr.h
@@ -1,15 +1,17 @@
 #ifndef MBR_H
 #define MBR_H
 
+#include <stdint.h>
+
 struct MBREntry {
 	uint8_t  boot_indicator;
 	uint8_t  starting_head;
-	uint8_t  starting_sector   :  6;
-	uint8_t  starting_cylinder : 10;
+	uint16_t starting_sector   :  6;
+	uint16_t starting_cylinder : 10;
 	uint8_t  system_id;
 	uint8_t  ending_head;
-	uint8_t  ending_sector     :  6;
-	uint8_t  ending_cylinder   : 10;
+	uint16_t ending_sector     :  6;
+	uint16_t ending_cylinder   : 10;
 	uint32_t starting_lba;
 	uint32_t sector_count;
 } __attribute__((packed));
